Refactor: Move TileDescription to UI package
